# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: MIT Pokemon Automation API
  description: |
    API for the MIT Pokemon Automation and Gameplay Application.
    All endpoints require authentication via X-API-Key header.
  version: 1.0.0
  contact:
    name: MIT Pokemon Support
servers:
  - url: /api
    description: API routes
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: User-specific secret key obtained from dashboard
  schemas:
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
    Pokemon:
      type: object
      required:
        - id
        - pokemon_id
        - name
        - level
        - types
        - sr
      properties:
        id:
          type: string
          format: uuid
          description: Owned Pokemon record ID
        pokemon_id:
          type: integer
          description: Pokemon species ID
        name:
          type: string
          description: Pokemon species name
        level:
          type: integer
          minimum: 1
          maximum: 10
          description: Pokemon level (condensed 1-10)
        types:
          type: array
          items:
            type: string
          description: Pokemon types
        sr:
          type: number
          description: Strength Rating
        selected_moves:
          type: array
          items:
            type: string
          maxItems: 4
          description: Currently selected moves
        is_active:
          type: boolean
          description: Whether this is the active Pokemon
        sprite_url:
          type: string
          format: uri
          description: Pokemon sprite image URL
...
    Dashboard:
      type: object
      required:
        - trainer_name
        - active_pokemon
        - stats
      properties:
        trainer_name:
          type: string
        active_pokemon:
          $ref: '#/components/schemas/Pokemon'
        pokemon_count:
          type: integer
          description: Total Pokemon owned
        stats:
          type: object
          properties:
            money:
              type: integer
            items:
              type: object
              additionalProperties:
                type: integer
            battles_won:
              type: integer
            battles_lost:
              type: integer
            pokemon_captured:
              type: integer
        has_active_battle:
          type: boolean
    Move:
      type: object
      required:
        - id
        - name
        - type
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        description:
          type: string
    SecretKey:
      type: object
      required:
        - key
        - created_at
      properties:
        key:
          type: string
          description: The secret key (only shown once on generation)
        created_at:
          type: string
          format: date-time
paths:
  /dashboard:
    get:
      summary: Get dashboard data
      description: Returns aggregated dashboard data including active Pokemon, stats, and navigation info
      operationId: getDashboard
      tags:
        - Dashboard
      responses:
        '200':
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /secret-key:
    get:
      summary: Get secret key metadata
      description: Returns metadata about the user's secret key (not the key itself)
      operationId: getSecretKeyMeta
      tags:
        - Authentication
      responses:
        '200':
          description: Secret key metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  has_key:
                    type: boolean
                  created_at:
                    type: string
                    format: date-time
                  last_used_at:
                    type: string
                    format: date-time
        '401':
          description: Unauthorized

    post:
      summary: Generate or regenerate secret key
      description: |
        Generates a new secret key. If a key already exists, it is invalidated.
        The plaintext key is only returned once and cannot be retrieved again.
      operationId: generateSecretKey
      tags:
        - Authentication
      responses:
        '201':
          description: Secret key generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecretKey'
        '401':
          description: Unauthorized
...
  /pokecenter:
    get:
      summary: Get Pokemon collection
      description: Returns all Pokemon owned by the user
      operationId: getPokemonCollection
      tags:
        - Pokecenter
      responses:
        '200':
          description: Pokemon collection
          content:
            application/json:
              schema:
                type: object
                properties:
                  pokemon:
                    type: array
                    items:
                      $ref: '#/components/schemas/Pokemon'
                  active_pokemon_id:
                    type: string
                    format: uuid
        '401':
          description: Unauthorized
...
  /moves:
    get:
      summary: Get available moves for active Pokemon
      description: Returns all moves available to the user's active Pokemon
      operationId: getAvailableMoves
      tags:
        - Moves
      responses:
        '200':
          description: Available moves
          content:
            application/json:
              schema:
                type: object
                properties:
                  available_moves:
                    type: array
                    items:
                      $ref: '#/components/schemas/Move'
                  selected_moves:
                    type: array
                    items:
                      type: string
                    maxItems: 4
        '401':
          description: Unauthorized

    put:
      summary: Update selected moves
      description: Sets the 4 moves for the active Pokemon's moveset
      operationId: updateSelectedMoves
      tags:
        - Moves
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - moves
              properties:
                moves:
                  type: array
                  items:
                    type: string
                  minItems: 4
                  maxItems: 4
                  description: Array of 4 move IDs
      responses:
        '200':
          description: Moves updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  selected_moves:
                    type: array
                    items:
                      type: string
                    maxItems: 4
        '400':
          description: Bad request (e.g., invalid moves, wrong number of moves)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
  /trainer:
    get:
      summary: Get trainer profile
      description: Returns the authenticated user's trainer profile
      operationId: getTrainerProfile
      tags:
        - Trainer
      responses:
        '200':
          description: Trainer profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  name:
                    type: string
                  role:
                    type: string
                  created_at:
                    type: string
                    format: date-time
        '401':
          description: Unauthorized
...